chore(cloudflare-workers): extract package
