+++ packages/cloudflare-workers/webpack.config.js
@@ -2,9 +2,9 @@ const path = require('path')
 
 module.exports = {
   mode: 'production',
-  entry: path.join(__dirname, 'src', 'cloudflare-workers'),
+  entry: path.join(__dirname, 'src'),
   output: {
-    path: path.resolve(__dirname, 'dist-cf'),
+    path: path.resolve(__dirname, 'dist'),
     filename: 'index.js',
     publicPath: '/'
   },
