+++ packages/cloudflare/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@serlo/cloudflare",
+  "version": "0.0.0",
+  "dependencies": {
+    "cloudflare": "^2.0.0"
+  },
+  "private": true
+}
